refactor: store에 index로 export 정리
